Fix Products components & Category pages
