+++ src/pages/Home.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Banner from '../components/Banner';
-import Products from '../components/Products';
+import BannerMedium from '../components/BannerMedium';
 import './Home.module.scss';
 
 export default function Home() {
   return (
     <div className='container'>
       <Banner />
-      <Products />
+      <BannerMedium />
     </div>
   );
 }
