--- src/pages/ProductsLiving.module.scss
@@ -1,22 +1,0 @@
-.products {
-  color: $color-black;
-  display: flex;
-  justify-content: center;
-  padding-top: 50px;
-  padding-bottom: 150px;
-  .container {
-    width: $width-inner;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    .row {
-      width: 100%;
-      min-height: 50vh;
-      display: grid;
-      grid-template-columns: repeat(auto-fill, 180px);
-      flex-wrap: wrap;
-      justify-content: center;
-      gap: 20px;
-    }
-  }
-}
